[HandshakeSimulator] Header and source files
