+++ experimental/lib/Support/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_dynamatic_library(DynamaticExperimentalSupport
   CDGAnalysis.cpp
   StdProfiler.cpp
+  HandshakeSimulator.cpp
 
   LINK_LIBS PUBLIC
   MLIRIR
